add modal styles
