+++ app/components/GroupContainer/style.module.css
@@ -1,0 +1,9 @@
+.container {
+  background-color: rgb(58, 62, 110);
+  border-radius: 8px;
+  border: 4px solid white;
+  display: flex;
+  flex-direction: column;
+  padding: 10px;
+  gap: 8px;
+}
