+++ app/components/Modal/modal.module.css
@@ -21,3 +21,17 @@
   width: 36px;
   align-self: flex-end;
 }
+
+.openButton {
+  background: none;
+  border: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.3rem;
+  text-decoration: none;
+  color: black;
+}
+.openButton:hover {
+  cursor: pointer;
+}
