--- components/AddToGroup/Button/addToGroup.module.css
@@ -1,13 +1,0 @@
-.button {
-  background: none;
-  border: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.3rem;
-  text-decoration: none;
-  color: black;
-}
-.button:hover {
-  cursor: pointer;
-}
