+++ components/PosterButton/PosterButton.tsx
@@ -4,8 +4,8 @@ import { Movie, TV } from "@/utilities/interface";
 import { PropsWithChildren } from "react";
 import getPoster from "../../utilities/getPoster";
 import style from "./PosterButton.module.css";
-import useModal from "../Modal/useModal";
-import Modal from "../Modal/Modal";
+import useModal from "../../app/components/Modal/useModal";
+import Modal from "../../app/components/Modal/Modal";
 
 const PosterButton = ({ poster_path, title, name, children, id, media_type }: PropsWithChildren<Movie | TV>) => {
   const { isOpen, openModal, closeModal } = useModal();
