+++ package.json
@@ -27,6 +27,7 @@
     "nodemailer": "^6.9.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
     "swiper": "^9.1.0",
     "typescript": "4.9.5"
   },
